requirements.txt
grpcio
grpcio-tools
protobuf
black
mysqlclient
aqlalchemy
alembic
